+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Support `background_color_column` for `table`
 - Support `text_color_column` for `table`
 - Support `checkbox_text_color_column` for `table`
+- Support `image_text_color_column` for `table`
 
 ## 0.2.6